Use a JSON file for transpiler settings

+++ hardhat/hardhat-transpiler/tasks/transpile.ts
@@ -8,36 +8,31 @@ import { findAlreadyInitializable } from '@openzeppelin/upgrade-safe-transpiler/
 
 interface TranspileOptions {
   initializablePath?: string;
+  deleteOriginals?: boolean;
+  skipWithInit?: boolean;
   exclude?: string[];
   publicInitializers?: string[];
-  solcVersion?: string;
-  skipWithInit?: boolean;
   namespaced?: boolean;
   namespaceExclude?: string[];
   peerProject?: string;
 }
 
-const options: TranspileOptions = {
-  initializablePath: 'project/contracts/proxy/utils/Initializable.sol',
-  exclude: [
-    'project/contracts-exposed/**/*',
-    'project/contracts/mocks/**/*Proxy*.sol',
-    'project/contracts/proxy/**/*Proxy*.sol',
-    'project/contracts/proxy/beacon/UpgradeableBeacon.sol',
-  ],
-  publicInitializers: [
-    'project/contracts/access/manager/AccessManager.sol',
-    'project/contracts/finance/VestingWallet.sol',
-    'project/contracts/governance/TimelockController.sol',
-    'project/contracts/metatx/ERC2771Forwarder.sol',
-  ],
-  namespaced: true,
-  namespaceExclude: ['project/contracts/mocks/**/*'],
-  peerProject: '@openzeppelin/',
-};
+async function loadSettings(hre: HardhatRuntimeEnvironment, settings: string): Promise<TranspileOptions> {
+  const makeProjectPath = (p: string) =>
+    path.join('project', path.isAbsolute(p) ? path.relative(hre.config.paths.root, p) : p);
 
-export default async function (taskArguments: TranspileOptions = {}, hre: HardhatRuntimeEnvironment) {
-  // TODO Use taskArguments to override options
+  const options: TranspileOptions = await fs.readFile(settings, 'utf-8').then(JSON.parse);
+  options.initializablePath = options.initializablePath && makeProjectPath(options.initializablePath);
+  options.exclude = options.exclude?.map(makeProjectPath);
+  options.publicInitializers = options.publicInitializers?.map(makeProjectPath);
+  options.namespaceExclude = options.namespaceExclude?.map(makeProjectPath);
+
+  return options;
+}
+
+export default async function ({ settings }: { settings?: string }, hre: HardhatRuntimeEnvironment) {
+  assert(settings, 'Transpile settings file must be provided');
+  const options = await loadSettings(hre, settings);
 
   // Clean compilation (should result in a single build info)
   await hre.tasks.getTask('clean').run();
@@ -74,18 +69,23 @@ export default async function (taskArguments: TranspileOptions = {}, hre: Hardha
   );
 
   // Delete originals files
-  const keep = new Set(
-    ([] as string[])
-      .concat(
-        transpiled.map(t => t.path),
-        findAlreadyInitializable(output, options.initializablePath),
-      )
-      .map(p => path.join(hre.config.paths.root, p.replace(/^project\//, ''))),
-  );
-  await Promise.all(
-    Object.keys(output.sources)
-      .map(s => path.join(hre.config.paths.root, s.replace(/^project\//, '')))
-      .filter(p => !keep.has(p))
-      .map(p => fs.unlink(p).catch(() => {})),
-  );
+  if (options.deleteOriginals) {
+    const keep = new Set(
+      ([] as string[])
+        .concat(
+          transpiled.map(t => t.path),
+          findAlreadyInitializable(output, options.initializablePath),
+        )
+        .map(p => path.join(hre.config.paths.root, p.replace(/^project\//, ''))),
+    );
+    if (options.initializablePath && options.peerProject === undefined) {
+      keep.add(path.join(hre.config.paths.root, options.initializablePath.replace(/^project\//, '')));
+    }
+    await Promise.all(
+      Object.keys(output.sources)
+        .map(s => path.join(hre.config.paths.root, s.replace(/^project\//, '')))
+        .filter(p => !keep.has(p))
+        .map(p => fs.unlink(p).catch(() => {})),
+    );
+  }
 }
